requirements-kestra.txt
requests
duckdb
pandas
python-dotenv
google-cloud-storage
google-cloud-bigquery
pyarrow
